requirements.txt
# Django stack
Django>=4.2
djangorestframework>=3.14
channels>=4.0
daphne>=4.0

# Twilio + WebSocket
twilio>=9.6
webrtcvad>=2.0
redis>=5.0

# AI + Audio
openai>=1.93
openai-whisper>=20250625
torch>=2.7
tiktoken>=0.9.0
sarvamai>=0.1.12

# Audio processing
pydub>=0.25
soundfile>=0.13
ffmpeg-python>=0.2
numpy>=2.2
scipy>=1.16

# LangChain stack (optional prune)
langchain>=0.3.26
langchain-core>=0.3.68
langchain-community>=0.3.27
langchain-openai>=0.3.27
langchain-groq>=0.3.6
langsmith>=0.4.6
langchain-tavily>=0.2.9
langchain-text-splitters>=0.3.8
groq>=0.30.0

# HTTP/Async
httpx>=0.28
httpcore>=1.0
httpx-sse>=0.4
aiohttp>=3.12
aiohttp-retry>=2.9
aiofiles>=0.8

# Config + Serialization
pydantic>=2.11
pydantic-settings>=2.10
dataclasses-json>=0.6
marshmallow>=3.26
typing-inspect>=0.9
typing-extensions>=4.14
python-dotenv>=1.1

# Miscellaneous
orjson>=3.11
regex>=2024.11
tqdm>=4.67
yarl>=1.20
